Refactor tests as per the new layout.
