+++ src/plugins/intel_cpu/tests/functional/single_layer_tests/instances/common/random_uniform.cpp
@@ -1,0 +1,68 @@
+// Copyright (C) 2018-2023 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#include "single_layer_tests/classes/random_uniform.hpp"
+
+using namespace CPUTestUtils;
+using namespace ov::test;
+
+namespace CPULayerTestsDefinitions {
+namespace RandomUniform {
+
+static const std::vector<ElementType> shape_prc = {
+        ElementType::i32,
+        ElementType::i64
+};
+
+static const std::vector<ov::Shape> output_shapes = {
+        {500},
+        {4, 3, 210}
+};
+
+static const std::vector<uint64_t> global_seed = {
+        0, 8
+};
+
+static const std::vector<uint64_t> operational_seed = {
+        0, 3, 5
+};
+
+static const std::vector<std::tuple<double, double>> min_max = {
+        {0, 50},
+        {-50, 50},
+        {-50, 0}
+};
+
+INSTANTIATE_TEST_SUITE_P(smoke_Param, RandomUniformLayerTestCPU,
+        ::testing::Combine(
+                ::testing::ValuesIn(output_shapes),
+                ::testing::ValuesIn(min_max),
+                ::testing::ValuesIn(shape_prc),
+                ::testing::Values(ElementType::f32, ElementType::i32),
+                ::testing::ValuesIn(global_seed),
+                ::testing::ValuesIn(operational_seed),
+                ::testing::Values(false),
+                ::testing::Values(false),
+                ::testing::Values(false),
+                ::testing::Values(emptyCPUSpec),
+                ::testing::Values(empty_plugin_config)),
+        RandomUniformLayerTestCPU::getTestCaseName);
+
+INSTANTIATE_TEST_SUITE_P(smoke_ParamConst, RandomUniformLayerTestCPU,
+        ::testing::Combine(
+                ::testing::Values(output_shapes[0]),
+                ::testing::Values(min_max[0]),
+                ::testing::Values(ElementType::i32),
+                ::testing::Values(ElementType::f32),
+                ::testing::Values(1),
+                ::testing::Values(0),
+                ::testing::Values(true, false),
+                ::testing::Values(true, false),
+                ::testing::Values(true, false),
+                ::testing::Values(emptyCPUSpec),
+                ::testing::Values(empty_plugin_config)),
+        RandomUniformLayerTestCPU::getTestCaseName);
+
+} // namespace RandomUniform
+} // namespace CPULayerTestsDefinitions
